hook up create package repo filter
